+++ index.html
@@ -233,7 +233,7 @@
   </template>
 
   <!-- JS -->
-  <script src="./js/main.js"></script>
-  <script src="./js/functions.js"></script>
+  <script src="./js/main.js" type="module"></script>
+  <!-- <script src="./js/functions.js"></script> -->
 </body>
 </html>
